refactor: move syncPaginationWithQueryParams from features to shared/library
