+++ src/shared/library/sync-pagination-with-query-params.ts
